requirements.txt
# Telegram Bot
python-telegram-bot==20.7

# Vector Database & Embeddings
chromadb==0.4.15
sentence-transformers==2.2.2

# AI/ML Dependencies
transformers==4.30.2
huggingface-hub==0.24.6
torch>=2.0.0

# LLM Provider
groq==0.4.1

# Web Framework
flask==3.0.0
streamlit==1.28.0

# PDF Processing
PyPDF2==3.0.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Optional: For better performance
# accelerate==0.24.0
